Log to new file each day
Issue:
Logger wasnt logging to a new file each day

Why does the bug exist:
I didnt test properly when writing the code. When the server starts it
creates a file to log to. Due to the way that we were originally running
the server i.e. restarting the server to get around other issues. A new
file was created each day giving the appearance that logging was working
correctly. When we moved to a new server, we didnt have to restart the
server everyday effectively 'breaking' the logging

The fix:
use node scheduler to change where the logger logs to each day